+++ CHANGELOG.md
@@ -4,4 +4,5 @@ CHANGELOG
 dev-master
 ----------
 
-* bugfix [#229] Userland node type filtering does not work
+* bugfix #229 Userland node type filtering does not work
+* feature #245 Added NodeProcessor class which can be used by implementations to validate and process nodes.